Adjusted Camera; Added Main Menu Scene
I added a scene called "Main Menu" and added a basic layout for the buttons. They currently have no functionality, but exist and have the city background for now (which I'm sure will change). In the gameplay scene, I adjusted camera and player position so it wasn't viewing out of bounds.