Runaway Script altered
Guy now runs away properly can change the speed

Also now can detect whether the player is in range of the "food" as the tag calls it. The "food"/guy will start running away.

Bug that might be an issue
sometimes you will not be able to eat the guy