Added PauseMenu and Script
Added a Canvas containing the pause menu. Added PauseMenu script that pauses time functions, keeps track of if the game is paused; modified "Update()" in the Player Script so that input only registers when game is unpaused. Resume and Quit buttons function as intended; Options button can be wired to that menu when we develop it